# openapi.yaml
openapi: 3.0.3
info:
  title: Kanban System API
  description: 看板工作流管理系统 API
  version: 2.0.0
servers:
  - url: /api
    description: API Server
...
paths:
...
  /boards/{id}/lists:
    get:
      tags:
        - list
      summary: 获取看板的列表
      operationId: getLists
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/List'
    post:
      tags:
        - list
      summary: 创建列表
      operationId: createList
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateListRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/List'
  /lists/{id}:
    put:
      tags:
        - list
      summary: 更新列表
      operationId: updateList
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateListRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/List'
    delete:
      tags:
        - list
      summary: 删除列表
      operationId: deleteList
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'

  # 列自动指派接口
  /lists/{id}/auto-assignments:
    get:
      tags:
        - list
      summary: 获取列的自动指派配置
      operationId: getAutoAssignments
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ListAutoAssignment'
    post:
      tags:
        - list
      summary: 设置列的自动指派
      operationId: setAutoAssignment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetAutoAssignmentRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAutoAssignment'
    delete:
      tags:
        - list
      summary: 删除列的自动指派
      operationId: deleteAutoAssignment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'

  # 状态转移规则接口
...
  /boards/{id}/members:
    get:
      tags:
        - member
      summary: 获取看板成员
      operationId: getBoardMembers
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BoardMember'
    post:
      tags:
        - member
      summary: 添加看板成员
      operationId: addBoardMember
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddBoardMemberRequest'
      responses:
        '201':
          description: 添加成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardMember'
  /boards/{id}/members/{user_id}:
    put:
      tags:
        - member
      summary: 更新成员角色
      operationId: updateBoardMemberRole
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMemberRoleRequest'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardMember'
    delete:
      tags:
        - member
      summary: 移除看板成员
      operationId: removeBoardMember
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'

  # 活动日志接口
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
          format: email
        nickname:
          type: string
        avatar:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - username
        - email
...
    List:
      type: object
      properties:
        id:
          type: integer
        board_id:
          type: integer
        title:
          type: string
        position:
          type: integer
        wip_limit:
          type: integer
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        cards:
          type: array
          items:
            $ref: '#/components/schemas/Card'
      required:
        - id
        - board_id
        - title
        - position
...
    Card:
      type: object
      properties:
        id:
          type: integer
        list_id:
          type: integer
        swimlane_id:
          type: integer
          nullable: true
        title:
          type: string
        description:
          type: string
        position:
          type: integer
        due_date:
          type: string
          format: date-time
          nullable: true
        cover:
          type: string
        assignee_id:
          type: integer
          nullable: true
        assignee:
          $ref: '#/components/schemas/User'
        completed_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        labels:
          type: array
          items:
            $ref: '#/components/schemas/Label'
        comments:
          type: array
          items:
            $ref: '#/components/schemas/Comment'
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
        checklist_items:
          type: array
          items:
            $ref: '#/components/schemas/ChecklistItem'
        checklist_progress:
          type: object
          properties:
            total:
              type: integer
            completed:
              type: integer
      required:
        - id
        - list_id
        - title
        - position
    ChecklistItem:
      type: object
      properties:
        id:
          type: integer
        card_id:
          type: integer
        content:
          type: string
        completed:
          type: boolean
        position:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - card_id
        - content
        - completed
        - position
    Label:
      type: object
      properties:
        id:
          type: integer
        board_id:
          type: integer
        name:
          type: string
        color:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - board_id
        - name
        - color
    Comment:
      type: object
      properties:
        id:
          type: integer
        card_id:
          type: integer
        user_id:
          type: integer
        user:
          $ref: '#/components/schemas/User'
        content:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - card_id
        - user_id
        - content
    Attachment:
      type: object
      properties:
        id:
          type: integer
        card_id:
          type: integer
        file_name:
          type: string
        file_url:
          type: string
        file_size:
          type: integer
        mime_type:
          type: string
        created_at:
          type: string
          format: date-time
      required:
        - id
        - card_id
        - file_name
        - file_url
    BoardMember:
      type: object
      properties:
        id:
          type: integer
        board_id:
          type: integer
        user_id:
          type: integer
        user:
          $ref: '#/components/schemas/User'
        role:
          type: string
          enum:
            - owner
            - admin
            - member
            - observer
        joined_at:
          type: string
          format: date-time
      required:
        - id
        - board_id
        - user_id
        - role
...
    ListAutoAssignment:
      type: object
      properties:
        id:
          type: integer
        list_id:
          type: integer
        user_id:
          type: integer
        user:
          $ref: '#/components/schemas/User'
      required:
        - id
        - list_id
        - user_id
...
    CreateListRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 100
        wip_limit:
          type: integer
          nullable: true
      required:
        - title
    UpdateListRequest:
      type: object
      properties:
        title:
          type: string
          maxLength: 100
        position:
          type: integer
        wip_limit:
          type: integer
          nullable: true
...
    AddBoardMemberRequest:
      type: object
      properties:
        user_id:
          type: integer
        role:
          type: string
          enum:
            - admin
            - member
            - observer
          default: member
      required:
        - user_id
    UpdateMemberRoleRequest:
      type: object
      properties:
        role:
          type: string
          enum:
            - admin
            - member
            - observer
      required:
        - role
    SetAutoAssignmentRequest:
      type: object
      properties:
        user_id:
          type: integer
      required:
        - user_id
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message